requirements.txt
fastapi
uvicorn
streamlit
requests
openai
langchain
python-dotenv
pydantic
numpy
torch
transformers
librosa
chromadb
sentence-transformers